move georeference app, use app.task not shared_task
