+++ ohmg/celeryapp.py
@@ -12,7 +12,6 @@ app = Celery('ohmg')
 app.config_from_object('django.conf:settings', namespace="CELERY")
 app.autodiscover_tasks()
 
-
 @app.task(bind=True)
 def debug_task(self):
     print("Request: {!r}".format(self.request))
