+++ ohmg/georeference/iiif/urls.py
